Re-organize file structure and naming
