+++ src/components/Layout/OptionsRow.tsx
@@ -1,19 +1,17 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
-import RadioButtonsGroup from '../inputs/radioButtons';
-import { CheckboxesGroup } from '../inputs/checkboxes';
-import MeasurementPeriodDatePicker from '../inputs/datePicker';
+import { CalculationOptionsButtons, MeasurementPeriodDatePicker, OutputTypeButtons } from '../CalculationOptions';
 
 export default function OptionsRow() {
   return (
     <React.Fragment>
       <Grid item xs={4}>
         <h3>Output Type:</h3>
-        <RadioButtonsGroup />
+        <OutputTypeButtons />
       </Grid>
       <Grid item xs={4}>
         <h3>Calculation Options:</h3>
-        <CheckboxesGroup />
+        <CalculationOptionsButtons />
       </Grid>
       <Grid item xs={4}>
         <MeasurementPeriodDatePicker />
