+++ src/components/Layout/index.ts
@@ -1,0 +1,2 @@
+export { default as DataImportRow } from './DataImportRow';
+export { default as OptionsRow } from './OptionsRow';
