+++ src/components/Utils/index.ts
@@ -1,0 +1,1 @@
+export { default as Loading } from './Loading';
