+++ src/index.tsx
@@ -1,10 +1,9 @@
-import React, { Suspense } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import { RecoilRoot } from 'recoil';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Loading from './components/Loading';
 
 ReactDOM.render(
   <React.StrictMode>
